% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_ewma}
\alias{arl_ewma}
\title{Calibration Parameters for Exponentially Weighted Moving Average (EWMA) statistic}
\usage{
arl_ewma(
  x,
  mu,
  sigma2,
  lambda,
  L = seq(1, 3, by = 0.25),
  max.rl = 110,
  rl = 100,
  scenario = 0,
  nboot = 1000,
  ncores = 5,
  seed = 1234
)
}
\arguments{
\item{x}{A numeric vector of individual observations obtained from the process.
In some cases, \code{x} can also indicate sample averages according to a specified sampling plan.}

\item{mu}{double; process mean. Also known as the target mean or the average of the historical data.}

\item{sigma2}{double; estimate of the process variance.}

\item{lambda}{numeric; a weighting factor between 0 and 1 that determines how much emphasis is placed on recent versus historical data in EWMA charts. Smaller λ values give more weight to recent observations, making the chart more sensitive to small process shifts.}

\item{L}{A numeric value or vector. \code{L} is a control limit factor; represents
the width of the control limits in terms of multiples of the standard deviation.}

\item{max.rl}{integer; Maximum Run Length. It refers to the expected maximum number of observations in the experiment.}

\item{rl}{integer; Run Length of interest. It refers to the target number of observations to calculate either the percentage of false alarms or true alarms. See Details for more information.}

\item{scenario}{A logical evaluating to 0 or 1 indicating whether the process is under control \code{scenario} = 0, or the process out of control \code{scenario} = 1.}

\item{nboot}{integer; number of simulations to generate the process replicates.}

\item{ncores}{integer; number of CPU cores to use for parallel processing.}

\item{seed}{integer; to initialize the Random Number Generator to a known state.}
}
\value{
Returns a data frame containing the Operating Characteristics for EWMA charts.
The data frame includes the following columns:
\itemize{
\item \code{L}: Control Limit Factor.
\item \code{ARL}: Average Run Length. See Details for more information about ARL0 and ARL1.
\item \code{FP}: False Positive rate, representing the frequency of false alarms when the process is under control.
\item \code{TP}: True Positive rate, indicating the rate of correctly detecting a shift when the process is actually out of control.
}
}
\description{
\code{arl_ewma()} returns estimates of the Operating Characteristics for the
the EWMA quality control charts through
bootstrap sampling.
}
\details{
The Run Length of Interest; \code{rl} is the
same as the maximum run length when the process is under control because the summary statistic is the number of
false positive (FP) alarms across the entire duration of the experiment.
When the process is out-of-control, it is critical to detect true alarms quickly.
Therefore, the run length of interest is a small fraction of the maximum run
length because the summary statistic is the number of true positive (TP) alarms
in short period of time within the experiment.

\code{ARL} is the expected Run Length \code{rl} generated by \code{arl_ewma()}.
When the process is under control \code{scenario = 0}, Average Run Length is denoted as
\code{ARL0} and \code{FP} is returned. If the process is out of control
\code{scenario = 1}, Average Run Length is used as \code{ARL1} and \code{TP} is
returned.

\code{ARL0} and \code{ARL1} refer to the expected number of observations to
detect an out-of-control condition. When the process is under control, a high \code{ARL0} is
preferred as it demonstrates that the chart does not react excessively to regular
variations within process limits, thus minimizing false alarms. Conversely, a low
\code{ARL1} is desirable because it signifies that the chart is adequately sensitive,
enabling quick detection and response to actual deviations in the process. This
increases the true positive rate, hence the power of the quality control charts.
}
\examples{
x <- rnorm(100)
lambda = 0.2
max.rl = 110
aux = control_parm_ewma(x, lambda, max.rl)

## process under control
arl_ewma(x, mu = aux$mu, aux$sigma2, lambda = aux$lambda, 
         L = seq(1, 3, by = 0.25), 
         max.rl = 110, rl = 100, 
         scenario = 0, nboot = 100, 
         ncores = 1, seed = 1234)
         
## process out of control (shift in the mean by 0.5 standard deviation)
x <- rnorm(200) 
lambda = 0.2
max.rl = 110
x.shift = x + 0.5*(sqrt(var(x))) 
aux = control_parm_ewma(x, lambda, max.rl)
arl_ewma(x = x.shift, mu = aux$mu, aux$sigma2, lambda, 
         L = seq(1, 3, by = 0.25), 
         max.rl = 110, rl = 100, 
         scenario = 1, nboot = 100, 
         ncores = 1, seed = 1234)
}
