% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_parameters.R
\name{control_parm_ewma}
\alias{control_parm_ewma}
\title{Control parameters for the Exponential Weighted Moving Average (EWMA)}
\usage{
control_parm_ewma(x, lambda, max.rl)
}
\arguments{
\item{x}{A numeric vector of individual observations obtained from the process.
In some cases, \code{x} can also indicate sample averages according to a specified sampling plan.}

\item{lambda}{numeric; a weighting factor between 0 and 1 that determines how much emphasis is placed on recent versus historical data in EWMA charts. Smaller λ values give more weight to recent observations, making the chart more sensitive to small process shifts.}

\item{max.rl}{integer; Maximum Run Length. It refers to the expected maximum number of observations in the experiment.}
}
\value{
A list containing the following components:
\itemize{
\item \code{mu}: The process mean. Also known as the target mean or the average of the historical data \code{x}.
\item \code{sigma2}: The variance of EWMA statistic.
\item \code{lambda}: the weighting factor of EWMA charts.
}
}
\description{
\code{control_parm_ewma} computes the essential parameters required for generating
the Operating Characteristics of EWMA quality control charts. It includes the process
mean and the variance of EWMA statistic when the process is under control.
}
\details{
The variance of the EWMA statistic \code{sigma2} is given by:
\deqn{\sigma^2(Z_i) = \left[(1-(1-\lambda)^{2i})\frac{{\lambda}}{{2-\lambda}}\right]\sigma^2_X}
The Process Mean \code{mu} is the average of the individual observations obtained from \code{X}:
\deqn{\bar{X} = \frac{\sum_{j=0}^{n}X_j}{n}}
}
\examples{
x <- rnorm(100)
lambda <- 0.2
max.rl <- 500
control_parm_ewma(x, lambda, max.rl)

}
\seealso{
\code{\link{arl_ewma}}, \code{\link{ewma_statistic}},
\code{\link{lower_limit_ewma}}, \code{\link{upper_limit_ewma}}
}
