% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ewma_statistic_C}
\alias{ewma_statistic_C}
\title{Title}
\usage{
ewma_statistic_C(x, t, lambda, x0)
}
\arguments{
\item{x}{A dataframe}

\item{t}{A sequence of numeric values of same length of Maximum ARL}

\item{lambda}{An integer}

\item{x0}{Mean value an integer}
}
\value{
A vector of EWMA
}
\description{
Title
}
\examples{
ewma_statistic_C(x = 1:50, t = 1:30, lambda = 0.15, x0 = 2)
}
