% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ma_statistic}
\alias{ma_statistic}
\title{Moving Average statistic}
\usage{
ma_statistic(x, t, omega)
}
\arguments{
\item{x}{A numeric vector of individual observations obtained from the process.
In some cases, \code{x} can also indicate sample averages according to a specified sampling plan.}

\item{t}{An integer vector; representing a sequence from 1 to
n, where n denotes the number of observations used in the Moving Average.}

\item{omega}{integer; refers to the the window size for MA charts that determines how many consecutive observations are weighted equally. For example, with Omega = 10, each of the last 10 observations receives equal weight of 0.1 (⅒). This parameter directly impacts the chart’s smoothing behavior and sensitivity to process changes.}
}
\value{
A numeric vector of Moving average statistics.
}
\description{
\code{ma_statistic()} calculates moving averages for a provided time series vector,
using a defined window size.
}
\details{
The moving average statistic is given by:
\deqn{Z_i = \left\{
\begin{array}{ll}
\frac{1}{\omega} \sum_{j=t[i]-\omega+1}^{t[i]} x_j & t[i] \geq \omega \\\\
\frac{1}{t[i]} \sum_{j=1}^{t[i]} x_j & t[i] < \omega
\end{array}
\right.}

where \code{t[i]} is the observation index at time \code{i}, \eqn{\omega} is the weighting factor
\code{omega}, also known as the window size, on which averages are calculated,
and \code{x_j} is the individual observation of the process \code{x} at time \code{j}.
}
\examples{
x = rnorm(100)
ma_values <- ma_statistic(x, t = 1:length(x), omega = 10)
print(ma_values)

}
\seealso{
\code{\link{arl_ma}}, \code{\link{lower_limit_ma}},
\code{\link{upper_limit_ma}}, \code{\link{ewma_statistic}}
}
