% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ma_statistic_C}
\alias{ma_statistic_C}
\title{Title}
\usage{
ma_statistic_C(x, t, omega)
}
\arguments{
\item{x}{A dataframe}

\item{t}{A sequence of numeric values of same length of Maximum ARL}

\item{omega}{An integer}
}
\value{
A vector of MA
}
\description{
Title
}
\examples{
ma_statistic_C(x = 1:50, t = 1:30, omega = 0.15)
}
