% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arl.R
\name{arl_ma}
\alias{arl_ma}
\title{Calibration parameters for MA quality control charts}
\usage{
arl_ma(
  x,
  mu,
  sigma2,
  omega,
  L = seq(1, 3, by = 0.25),
  max.arl = 10000,
  arl = 100,
  scenario = 0,
  nboot = 100,
  ncores = 1,
  seed = 1234
)
}
\arguments{
\item{x}{A numeric vector containing the data points for which to calculate the MA.}

\item{mu}{double; process mean. Also known as the target mean or the average of the historical data.}

\item{sigma2}{double: estimate of the process variance.}

\item{omega}{integer; the weighting factor of MA charts.}

\item{L}{integer: the control limit factor; represents the width of the control limits in terms of multiples of the standard deviation.}

\item{max.arl}{integer; Maximum Average Run Length.}

\item{arl}{integer; Average Run Length.}

\item{scenario}{A logical evaluating to TRUE or FALSE indicating whether the process is under control \code{scenario} = 0, or the process out of control \code{scenario} = 1.}

\item{nboot}{Number of iterations to calculate the MA statistics}

\item{ncores}{integer; the number of CPU cores to use for parallel processing.}

\item{seed}{integer; to initialize the Random Number Generator to a known state.}
}
\value{
This function produces a data frame for calibration, 
which includes the control limit factor \code{L}, Average Run Length \code{ARL}, 
and either the False Positive rate (when the process is in control) 
or the True Negative rate (when the process is out of control).
}
\description{
This function produces a data frame for the calibration of Moving Average 
charts, which includes the control limit factor \code{L}, 
Average Run Length \code{ARL}, and either the False Positive rate 
(when the process is in control) or the True Negative rate 
(when the process is out of control).
}
