% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_parameters.R
\name{control_parm_ewma}
\alias{control_parm_ewma}
\title{Control parameters for the Exponential Weighted Moving Average}
\usage{
control_parm_ewma(x, lambda, max.arl)
}
\arguments{
\item{x}{A numeric vector containing the data points for which to calculate the EWMA.}

\item{lambda}{A numeric value between 0 and 1 inclusive that indicates the smoothing constant of EWMA charts.
It determines the weight given to recent data points. A smaller \code{lambda} gives more weight to recent observations,
making the chart more sensitive to shifts.}

\item{max.arl}{integer; Maximum Average Run Length.}
}
\value{
Returns mean and variance of the sample
}
\description{
Returns a list of values containing the process mean, the 
estimated variance of the Exponential Weighted Moving Average when the process is 
under control, and the value of lambda.
}
