% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_parameters.R
\name{control_parm_ma}
\alias{control_parm_ma}
\title{Control parameters for the Moving Average}
\usage{
control_parm_ma(x, omega)
}
\arguments{
\item{x}{A numeric vector containing the data points for which to calculate the MA.}

\item{omega}{An integer}
}
\value{
Returns mean and variance of the sample
}
\description{
Returns a list of values containing the target mean of the sample, the variance of 
the Moving Average when the process is under control, and the value of omega.
}
