% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ewma_statistic}
\alias{ewma_statistic}
\title{Exponentially Weighted Moving Average statistic}
\usage{
ewma_statistic(x, t, lambda, x0)
}
\arguments{
\item{x}{A numeric vector containing the data points for which to calculate the Exponentially Weighted Moving Averages.}

\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the number of observations used in the Exponential 
Weighted Moving Average.}

\item{lambda}{A numeric value between 0 and 1 inclusive that indicates the weighting factor of EWMA charts.
It determines the weight given to recent data points. A smaller \code{lambda} gives more weight to recent observations,
making the chart more sensitive to shifts.}

\item{x0}{double; process mean. Also known as the target mean or the average of the historical data.}
}
\value{
A numeric vector of Exponential Weighted Moving average statistic.
}
\description{
Exponentially Weighted Moving Average statistic
}
\examples{
ewma_statistic(x = 1:50, t = 1:30, lambda = 0.15, x0 = 2)
}
