% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ewma_statistic}
\alias{ewma_statistic}
\title{Exponentially Weighted Moving Average statistic}
\usage{
ewma_statistic(x, t, lambda, x0)
}
\arguments{
\item{x}{A numeric vector of individual observations obtained from the process.
In some cases, \code{x} can also indicate sample averages according to a specified sampling plan.}

\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the number of observations used in the Exponential 
Weighted Moving Average. \code{t} is denoted by \code{i} in the EWMA statistic formula. 
See Details for more information.}

\item{lambda}{A numeric value between 0 and 1 inclusive that indicates the weighting factor of EWMA charts.
It determines the weight given to recent data points. A smaller \code{lambda} gives more weight to recent observations,
making the chart more sensitive to shifts.}

\item{x0}{double; process mean. Also known as the target mean or the average of the 
historical data. \code{x0} is denoted by \code{Z_0} in the EWMA statistic formula. 
See Details for more information.}
}
\value{
A numeric vector of Exponential Weighted Moving average statistic.
}
\description{
\code{ewma_statistic()} computes exponentially weighted moving averages for a given 
time series, assigning greater importance to recent observations while still considering 
older data points.
}
\details{
The EWMA statistic is given by:
\deqn{Z_i = \lambda \sum_{j=0}^{i-1} (1-\lambda)X_{i-j} + (1-\lambda)^i Z_0}
}
\examples{
x = rnorm(100)
ewma_values <- ewma_statistic(x, t = 1:length(x), lambda = 0.15, x0 = mean(x))
print(ewma_values)

}
\seealso{
\code{\link{arl_ewma}}, \code{\link{lower_limit_ewma}}, 
\code{\link{upper_limit_ewma}}, \code{\link{ma_statistic}}
}
