% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ewma_statistic}
\alias{ewma_statistic}
\title{Exponentially Weighted Moving Average statistic}
\usage{
ewma_statistic(x, t, lambda, x0)
}
\arguments{
\item{x0}{}
}
\value{

}
\description{
Exponentially Weighted Moving Average statistic
}
