% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{lower_limit_ma}
\alias{lower_limit_ma}
\title{Lower Control Limits (LCL) for Moving Average}
\usage{
lower_limit_ma(t, omega, mu, sigma2, L)
}
\arguments{
\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the number of observations used in the Moving Average.}

\item{omega}{integer; the weighting factor of MA charts.}

\item{mu}{double; process mean. Also known as the target mean or the average of the historical data.}

\item{sigma2}{double: estimate of the process variance.}

\item{L}{integer: the control limit factor; represents the width of the control 
limits in terms of multiples of the standard deviation.}
}
\value{
MA Lower Limit
}
\description{
Lower Control Limits (LCL) for Moving Average
}
\examples{
lower_limit_ma(t = 1:30, omega =10, mu = 2, sigma2 = 4, L = 1)

}
