% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ma_statistic}
\alias{ma_statistic}
\title{Moving Average statistic}
\usage{
ma_statistic(x, t, omega)
}
\arguments{
\item{x}{A numeric vector}

\item{t}{An integer vector}

\item{omega}{An integer value}
}
\value{
A vector of MA
}
\description{
Moving Average statistic
}
\examples{
ma_statistic(x = 1:50, t = 1:30, omega = 0.15)
}
