% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.R
\name{ma_statistic}
\alias{ma_statistic}
\title{Moving Average statistic}
\usage{
ma_statistic(x, t, omega)
}
\arguments{
\item{x}{A numeric vector containing the data points for which to calculate the Moving Averages.}

\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the number of observations used in the Exponential 
Weighted Moving Average.}

\item{omega}{integer; the weighting factor of MA charts.}
}
\value{
A numeric vector of Moving average statistic.
}
\description{
Moving Average statistic
}
\examples{
ma_statistic(x = 1:50, t = 1:30, omega = 0.15)
}
