\name{qclc}
\alias{qclc}
\docType{package}
\title{
  Quality Control Limit Charts
}
\description{
  Control charts for Moving Average (MA) and Exponentially Weighted Moving Average (EWMA), supporting the monitoring of both in-control and out-of-control processes.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{

Author: Marcio Augusto Diniz <marcio.diniz@cshs.org>
        Karni Bedirian <kb3896@nyu.edu>

Maintainer: Marcio Augusto Diniz <marcio.diniz@cshs.org>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
