% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{upper_limit_ewma}
\alias{upper_limit_ewma}
\title{Upper Control Limits (UCL) for Exponentially Weighted Moving Average}
\usage{
upper_limit_ewma(t, mu, sigma2, L)
}
\arguments{
\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the sample size used for calculating the Exponentially 
Weighted Moving Average (EWMA) statistic.}

\item{mu}{double; target mean.}

\item{sigma2}{double: variance of the EWMA statistic.}

\item{L}{integer: the control limit factor; represents the width of the control limits in terms of multiples of the standard deviation.
In EWMA charts, control limits are set based on the standard deviation of the EWMA statistic, which incorporates the smoothing constant 
\code{lambda}. \code{L} determines how wide the control limits are set from the center line.}
}
\value{
EWMA Upper Limit
}
\description{
Upper Control Limits (UCL) for Exponentially Weighted Moving Average
}
\examples{
upper_limit_ewma (t = 1:30, mu = 2, sigma2 = 4, L = 1)

}
