% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{upper_limit_ma}
\alias{upper_limit_ma}
\title{Upper Control Limits (UCL) for Moving Average}
\usage{
upper_limit_ma(t, omega, mu, sigma2, L)
}
\arguments{
\item{t}{An integer vector; representing a sequence from 1 to 
n, where n denotes the sample size used for calculating the Moving Average 
(MA) statistic.}

\item{omega}{An integer value}

\item{mu}{double; target mean.}

\item{sigma2}{double: variance of the MA statistic.}

\item{L}{integer: the control limit factor; represents the width of the control 
limits in terms of multiples of the standard deviation.}
}
\value{
MA Upper Limit
}
\description{
Upper Control Limits (UCL) for Moving Average
}
\examples{
upper_limit_ma(t = 1:30, omega =10, mu = 2, sigma2 = 4, L = 1)

}
